Make KeyTransformer and SHA256KeyHasher multiplatform
